# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "My API",
        "description": "REST API for the sonotrak application."
    },
...
    "host": "localhost:8080",
    "basePath": "/",
    "schemes": [
        "http"
    ],
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "definitions": {
...
        "agency": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "agency_name": {
                    "type": "string"
                },
                "agency_url": {
                    "type": "string"
                },
                "agency_timezone": {
                    "type": "string"
                },
                "agency_phone": {
                    "type": "string"
                },
                "agency_lang": {
                    "type": "string"
                }
            }
        },
        "user": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "email": {
                    "type": "string",
                    "format": "email"
                },
                "administrator": {
                    "type": "boolean"
                },
                "Agency": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "route": {
            "type": "object",
            "properties": {
                "agency_id": {
                    "type": "string"
                },
                "route_short_name": {
                    "type": "string"
                },
                "route_long_name": {
                    "type": "string"
                },
                "route_desc": {
                    "type": "string"
                },
                "route_type": {
                    "type": "integer"
                },
                "route_url": {
                    "type": "string"
                },
                "route_color": {
                    "type": "string"
                },
                "route_text_color": {
                    "type": "string"
                }
            }
        },
...
        "vehicle": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "uniqueId": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "latitude": {
                    "type": "string"
                },
                "longitude": {
                    "type": "string"
                },
                "fixTime": {
                    "type": "string"
                },
                "speed": {
                    "type": "integer"
                },
                "course": {
                    "type": "integer"
                },
                "positionId": {
                    "type": "string"
                }
            }
        }
    },
    "paths": {
...
        "/api/user/{id}": {
            "put": {
                "summary": "Update an existing user",
                "description": "Updates an existing user with new information",
                "tags": [
                    "User"
                ],
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "user",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/user"
                        }
                    },
                    "401": {
                        "description": "Unauthenticated!"
                    }
                }
            },
            "get": {
                "summary": "get agency by id ",
                "description": "Updates an existing agency with new information",
                "tags": [
                    "User"
                ],
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/agency"
                        }
                    },
                    "401": {
                        "description": "Unauthenticated!"
                    }
                }
            },
            "delete": {
                "summary": "delete user by id ",
                "description": "delete an existing user",
                "tags": [
                    "User"
                ],
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/user"
                        }
                    },
                    "401": {
                        "description": "Unauthenticated!"
                    }
                }
            }
        },
...
        "/api/route/{id}": {
            "get": {
                "summary": "get route by id ",
                "description": "Updates an existing route with new information",
                "tags": [
                    "Route"
                ],
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/route"
                        }
                    },
                    "401": {
                        "description": "Unauthenticated!"
                    }
                }
            },
            "put": {
                "summary": "Update an existing route",
                "description": "Updates an existing route with new information",
                "tags": [
                    "Route"
                ],
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "user",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/route"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/route"
                        }
                    },
                    "401": {
                        "description": "Unauthenticated!"
                    }
                }
            },
            "delete": {
                "summary": "delete route by id ",
                "description": "delete an existing route",
                "tags": [
                    "Route"
                ],
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/route"
                        }
                    },
                    "401": {
                        "description": "Unauthenticated!"
                    }
                }
            }
        },
...
        "/api/vehicle/{id}": {
            "get": {
                "summary": "get vehicle by id ",
                "description": "Updates an existing vehicle with new information",
                "tags": [
                    "Vehicle"
                ],
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/vehicle"
                        }
                    },
                    "401": {
                        "description": "Unauthenticated!"
                    }
                }
            }
        },
...
    }
}